Code merge

+++ app/Http/Controllers/Company/RolesController.php
@@ -3,30 +3,27 @@
 namespace App\Http\Controllers\Company;
 
 use Exception;
-use App\Models\roles;
-use App\Models\CustomRole;
-use App\Rules\OnlyString;
+use App\Models\MenuRole;
 use Illuminate\Http\Request;
-use Illuminate\Validation\Rule;
 use App\Http\Controllers\Controller;
 use App\Http\Services\CustomRoleService;
 use Illuminate\Support\Facades\Validator;
 
 class RolesController extends Controller
 {
-    private $customRoleService; 
+    private $customRoleService;
     public function __construct(CustomRoleService $customRoleService)
     {
-            $this->customRoleService = $customRoleService;
+        $this->customRoleService = $customRoleService;
     }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $roles = $this->customRoleService->getRolesByCompanyID(auth()->guard('company')->user()->company_id);
-        
-        return view('company.roles_and_permission.roles.index')->with(['roles'=> $roles]);
+        $roles = $this->customRoleService->getRolesByCompanyID(Auth()->user()->company_id);
+
+        return view('company.roles_and_permission.roles.index')->with(['roles' => $roles]);
     }
 
     public function role_form()
@@ -38,8 +35,8 @@ class RolesController extends Controller
     {
 
         try {
-            $validator  = Validator::make($request->all(), [
-                'name' => ['required', 'string', 'unique:custom_roles,name'],
+            $validator = Validator::make($request->all(), [
+                'name' => ['required', 'string', 'unique:roles,name,NULL,id,company_id,' . auth()->user()->company_id],
             ]);
 
             if ($validator->fails()) {
@@ -47,19 +44,23 @@ class RolesController extends Controller
             }
 
             $data = $request->all();
-            $data['company_id'] = auth()->guard('company')->user()->company_id;
+            $data = $request->except(['_token']);
+            $data['company_id'] = Auth()->user()->company_id;
+            $data['created_by'] = Auth()->user()->id;
+            $data['category'] = 'custom';
+            
             if ($this->customRoleService->create($data)) {
 
                 return response()->json([
                     'message' => 'Role Created Successfully!',
-                    'data'   =>  view('company/roles_and_permission/roles/roles_list', [
-                        'roles' => $this->customRoleService->all()
+                    'data' => view('company/roles_and_permission/roles/roles_list', [
+                        'roles' => $this->customRoleService->getRolesByCompanyID(Auth()->user()->company_id)
                     ])->render()
                 ]);
             }
         } catch (Exception $e) {
 
-            return response()->json(['error' =>  $e->getMessage()], 400);
+            return response()->json(['error' => $e->getMessage()], 400);
         }
     }
 
@@ -68,8 +69,8 @@ class RolesController extends Controller
      */
     public function update(Request $request)
     {
-        $validator  = Validator::make($request->all(), [
-            'name' => ['required', 'string', new OnlyString, 'unique:states,name,' . $request->id],
+        $validator = Validator::make($request->all(), [
+            'name' => ['required', 'string', 'unique:roles,name,' . $request->id . ',id,company_id,' . auth()->user()->company_id],
         ]);
 
         if ($validator->fails()) {
@@ -78,12 +79,12 @@ class RolesController extends Controller
         $updateRole = $request->except(['_token', 'id']);
         $companyStatus = $this->customRoleService->updateDetails($updateRole, $request->id);
         if ($companyStatus) {
-            
+
             return response()->json(
                 [
                     'message' => 'Role Updated Successfully!',
-                    'data'   =>  view('company/roles_and_permission/roles/roles_list', [
-                        'roles' => $this->customRoleService->all()
+                    'data' => view('company/roles_and_permission/roles/roles_list', [
+                        'roles' => $this->customRoleService->getRolesByCompanyID(Auth()->user()->company_id)
                     ])->render()
                 ]
             );
@@ -98,8 +99,8 @@ class RolesController extends Controller
         if ($statusDetails) {
             return response()->json([
                 'message' => 'Role Status Updated Successfully!',
-                'data'   =>  view('company/roles_and_permission/roles/roles_list', [
-                    'roles' => $this->customRoleService->all()
+                'data' => view('company/roles_and_permission/roles/roles_list', [
+                    'roles' => $this->customRoleService->getRolesByCompanyID(Auth()->user()->company_id)
                 ])->render()
             ]);
         } else {
@@ -113,12 +114,27 @@ class RolesController extends Controller
     public function destroy(Request $request)
     {
         $id = $request->id;
+
+        $role = $this->customRoleService->getDetails($id);
+        
+        if(count($role->users)) {
+            return response()->json([
+                'status' => false,
+                'message' => 'This role is already assigned to users',
+                'data' => view('company/roles_and_permission/roles/roles_list', [
+                    'roles' => $this->customRoleService->getRolesByCompanyID(Auth()->user()->company_id)
+                ])->render()
+            ]);
+        }
+
         $data = $this->customRoleService->deleteDetails($id);
+        MenuRole::where('role_id', $request->role_id)->delete();
         if ($data) {
             return response()->json([
+                'status' => true,
                 'message' => 'Role Deleted Successfully!',
-                'data'   =>  view('company/roles_and_permission/roles/roles_list', [
-                    'roles' => $this->customRoleService->all()
+                'data' => view('company/roles_and_permission/roles/roles_list', [
+                    'roles' => $this->customRoleService->getRolesByCompanyID(Auth()->user()->company_id)
                 ])->render()
             ]);
         } else {
